confirm and cancel done

+++ src/main/java/com/siit/team24/OpenDoors/controller/AccommodationController.java
@@ -5,21 +5,20 @@ import com.siit.team24.OpenDoors.dto.reservation.AccommodationSeasonalRateDTO;
 import com.siit.team24.OpenDoors.dto.reservation.SeasonalRatesPricingDTO;
 import com.siit.team24.OpenDoors.dto.searchAndFilter.SearchAndFilterDTO;
 
+import com.siit.team24.OpenDoors.exception.ExistingReservationsException;
 import com.siit.team24.OpenDoors.model.*;
 
 import com.siit.team24.OpenDoors.model.enums.AccommodationType;
 import com.siit.team24.OpenDoors.model.enums.Amenity;
 import com.siit.team24.OpenDoors.service.AccommodationService;
 
+import com.siit.team24.OpenDoors.service.ReservationRequestService;
 import com.siit.team24.OpenDoors.service.user.UserService;
 import jakarta.persistence.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.*;
@@ -36,6 +35,9 @@ public class AccommodationController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private ReservationRequestService reservationRequestService;
 
   
     @GetMapping(value = "/all")
@@ -83,6 +85,10 @@ public class AccommodationController {
 //    @PreAuthorize("hasRole('HOST')")
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<Void> deleteAccommodation(@PathVariable Long id) {
+        if (!reservationRequestService.isAccommodationReadyForDelete(id))
+            throw new ExistingReservationsException();
+
+        reservationRequestService.denyAllFor(id);
         accommodationService.delete(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -112,7 +118,7 @@ public class AccommodationController {
 //    @PreAuthorize("hasRole('HOST')")
     @GetMapping(value = "/host/{hostId}")
     public ResponseEntity<Collection<AccommodationHostDTO>> getForHost(@PathVariable Long hostId) {
-        Collection<AccommodationHostDTO> accommodations = accommodationService.getForHost(hostId);
+        Collection<AccommodationHostDTO> accommodations = accommodationService.getDTOsForHost(hostId);
         return new ResponseEntity<>(accommodations, HttpStatus.OK);
     }
